backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
anthropic==0.40.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.151.0
sqlalchemy==2.0.36
asyncpg==0.30.0
redis==5.1.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
alembic==1.13.3
